+++ src/tools/voxedit/modules/voxedit-ui/SceneGraphPanel.cpp
@@ -44,7 +44,10 @@ void SceneGraphPanel::detailView(voxelformat::SceneGraphNode &node) {
 			ImGui::TextUnformatted(entry->key.c_str());
 			ImGui::TableNextColumn();
 			const core::String &id = core::string::format("##%i-%s", node.id(), entry->key.c_str());
-			ImGui::InputText(id.c_str(), &entry->value);
+			core::String value = entry->value;
+			if (ImGui::InputText(id.c_str(), &value, ImGuiInputTextFlags_EnterReturnsTrue | ImGuiInputTextFlags_AutoSelectAll)) {
+				sceneMgr().nodeSetProperty(node.id(), entry->key, value);
+			}
 			ImGui::TableNextColumn();
 			const core::String &deleteId = core::string::format(ICON_FA_TRASH "##%i-%s-delete", node.id(), entry->key.c_str());
 			if (ImGui::Button(deleteId.c_str())) {